reduced code repetition
the iswebkitversionlessorequalto is very nearly the same as iswebkitversiongreaterorequalto and happens to repeat quite a few calls, the proposed change is to remove the methods entirely and replace them with one call to get the userversion. The min and max versions can further be stored as numbers to reduce inefficiencies involved in calculating them.